create imageacquirer interface and package
Creating two implementations, docker-image-acquirer.go and
docker-container-acquirer.go
