+++ pkg/api/types.go
@@ -120,3 +120,9 @@ type Scanner interface {
 	// Name is the scanner's name
 	Name() string
 }
+
+// ImageAcquirer abstract getting an image and extracting it in a given directory
+type ImageAcquirer interface {
+	// Acquire gets the image from `source` and extract it in `dest` which is the first output
+	Acquire(source string) (string, docker.Image, ScanResult, FilesFilter, error)
+}
