+++ pkg/imageacquirer/test/dockercfg1
