+++ pkg/imageacquirer/test/dockercfg2
