+++ pkg/imageacquirer/test/dockercfg3
