+++ pkg/imageacquirer/test/passwordFile1
