+++ pkg/imageacquirer/util_test.go
@@ -1,0 +1,64 @@
+package imageacquirer
+
+import (
+	"testing"
+)
+
+func TestGetAuthConfigs(t *testing.T) {
+	goodNoAuth := AuthsOptions{}
+
+	goodTwoDockerCfg := AuthsOptions{}
+	goodTwoDockerCfg.DockerCfg.Values = []string{"test/dockercfg1", "test/dockercfg2"}
+
+	goodUserAndPass := AuthsOptions{}
+	goodUserAndPass.Username = "erez"
+	goodUserAndPass.PasswordFile = "test/passwordFile1"
+
+	badUserAndPass := AuthsOptions{}
+	badUserAndPass.Username = "erez"
+	badUserAndPass.PasswordFile = "test/nosuchfile"
+
+	badDockerCfgMissing := AuthsOptions{}
+	badDockerCfgMissing.DockerCfg.Values = []string{"test/dockercfg1", "test/nosuchfile"}
+
+	badDockerCfgWrong := AuthsOptions{}
+	badDockerCfgWrong.DockerCfg.Values = []string{"test/dockercfg1", "test/passwordFile1"}
+
+	badDockerCfgNoAuth := AuthsOptions{}
+	badDockerCfgNoAuth.DockerCfg.Values = []string{"test/dockercfg1", "test/dockercfg3"}
+
+	tests := map[string]struct {
+		opts          AuthsOptions
+		expectedAuths int
+		shouldFail    bool
+	}{
+		"two dockercfg":               {opts: goodTwoDockerCfg, expectedAuths: 3, shouldFail: false},
+		"username and passwordFile":   {opts: goodUserAndPass, expectedAuths: 1, shouldFail: false},
+		"two dockercfg, one missing":  {opts: badDockerCfgMissing, expectedAuths: 2, shouldFail: false},
+		"two dockercfg, one wrong":    {opts: badDockerCfgWrong, expectedAuths: 2, shouldFail: false},
+		"two dockercfg, no auth":      {opts: badDockerCfgNoAuth, expectedAuths: 2, shouldFail: false},
+		"password file doens't exist": {opts: badUserAndPass, expectedAuths: 1, shouldFail: true},
+		"no auths, default expected":  {opts: goodNoAuth, expectedAuths: 1, shouldFail: false},
+	}
+
+	for k, v := range tests {
+		auths, err := getAuthConfigs(v.opts)
+		if !v.shouldFail {
+			if err != nil {
+				t.Errorf("%s expected to succeed but received %v", k, err)
+			}
+			var authsLen int = 0
+			if auths != nil {
+				authsLen = len(auths.Configs)
+			}
+			if auths == nil || v.expectedAuths != authsLen {
+				t.Errorf("%s expected len to be %d but got %d from %v",
+					k, v.expectedAuths, authsLen, auths)
+			}
+		} else {
+			if err == nil {
+				t.Errorf("%s should have failed be it didn't", k)
+			}
+		}
+	}
+}
